+++ R/aaa.R
@@ -1,0 +1,4 @@
+on_load({
+  # Singletons
+  the <- new_environment()
+})