+++ R/utils.R
@@ -66,3 +66,10 @@ node_walk_replace <- function(node, old, new) {
     node <- node_cdr(node)
   }
 }
+
+# Compat for R < 4.0.0
+tryInvokeRestart <- function(r, ...) {
+  if (!is_null(findRestart(r))) {
+    invokeRestart(r, ...)
+  }
+}